[UPDATE] Clean up printing OS name
