+++ engine/utility/public/log.hpp
@@ -5,6 +5,7 @@ namespace spdlog
 {
 // After this is called, the logs are written to a file, instead of the console.
 void StartWritingToFile();
+void PrintOSName();
 }
 
 namespace bblog = spdlog;
